Merged PR 206365: Refactored style naming and created UIFabric theme
##Style Naming
The naming standard for styles in Android is Pascal case with dot notation. Dot notation works when inheriting from another style so parent styles have been added. Where an exact one to one match exists between UIFabric styles and Android styles the style is not adjusted. Inheriting styles has the advantage of leveraging font weight, size and text colors of parent styles.

Additionally "TextAppearance" is appended to text styles as is the standard implemented in Outlook and by other Android developers.

##Theme.UIFabric
`Theme.UIFabric` has been added to declare `textColorPrimary` as well as `colorPrimary` and `colorPrimaryDark` which are used in the ToolBar and Statusbar colors, now using Fabric colors.

`textColorPrimary` is added to the application theme enabling utilization of this text color by many of the parent TextAppearance styles. As a result text color is removed many of the TextViews in **_fragment_typography.xml_**.

This theme is extended in the Demo app. Where Fabric Android is consumed this Theme can be used as a parent where properties can be added and others overriden to suit the theme of the consumer App.

##CalendarDayView Colors
Two styles, `TextAppearance.UIFabric.CalendarDay` and  `TextAppearance.UIFabric.CalendarDay2` derive from the base font styles, `TextAppearance.UIFabric.Subheading` and `TextAppearance.UIFabric.Subheading2` as they all share font weight and size. However, the colors are being handled in **_calendar_week_day_text_selector.xml_** which necessitates the parent color be overriden with `@null `

##Primary Dark
Consulted with @<Jack Cowan> to add `uifabric_primaryDark`.

Related work items: #658677