Merged PR 251134: Splash / Launch screen for demo app
This PR introduces a launch screen xml to the demo that serves as a template that other apps can use to get a launch screen that matches the Fabric design toolkit.

To test the launch screen without it disappearing, you can comment out everything in `onCreate` in `DemoListActivity` except `super.onCreate(savedInstanceState)`.

| Portrait | Landscape |
| ------------- |:-------------:|
| ![Screenshot_1553207368.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/251134/attachments/Screenshot_1553207368.png) | ![Screenshot_1553207468.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/251134/attachments/Screenshot_1553207468.png) |

Related work items: #701127