Merged PR 257132: README updates for GitHub
README updates for public GitHub, based on a variety of existing comparable repos (listed in the work item). Should have all the basics of what it is and how to use it.

No screenshots of controls yet, we should decide how we want to host them. Adding them via GitHub itself might actually be easier once we're doing PRs on there.

Related work items: #711210