Merged PR 266853: CalendarView 2 Row Bug Fix
Repro:
When the CalendarView is set to any mode other than `DisplayMode.FULL_MODE` the current week is at the beginning of the month, and the `animateResize` flag is set to true, the CalendarView scrolls itself to the beginning of the grid, which is 100 years into the past.

Fix
This fix ensures that the current or selected date is shown by setting a default date, which is `LocalDate.now()`, and calling a method that ensures the date is visible when the resize animation is finished.

FYI: Removed code that wasn't needed or had no real use

Test:
In `CalendarViewActivity`, insert `calendar_view.setDisplayMode(CalendarView.DisplayMode.PREVIEW_MODE)` into the `onCreate` method.

Related work items: #718403