Merged PR 245734: New ListItemView
This PR introduces a generic `ListItemView` that is designed to be used as content for rows in `RecyclerView`s, `ListView`s, etc. The design toolkit has many variants, and this PR only tackles some of them. I have a `TODO` list at the top of the `ListItemView` class documenting remaining variants to account for.

Here is a list of variants included in this PR:
- Basic three-line, two-line, and single-line lists
- Basic sub header
- Text wrapping
- Dense list items
- Custom views for adding icons and avatars on the left
- Custom accessory views for icons and text, etc, on the right
- Row divider / decorator

![update list gif.gif](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/245734/attachments/update%20list%20gif.gif)

Related work items: #700819