Merged PR 281069: LargeHeader variant for ListItemView
We will need this variant for `PersonaView` when I update that control to extend `ListItemView`.

Added another `CustomViewSize`, `LARGE`, that will map to `AvatarSize.XXLarge`.  When that size is used, the text appearance and list item margins will adjust to the "Large Header" style. Currently there are no planned implementations for this style that do not include a custom view so I kept the api simple.

Currently the toolkit does not show the "Large Header" variant used in an actual full list, but I confirmed with Jack that using it as a regular list item is a possibility.

The Title+Footer variant is in the demo as a special request from Jack.

![Screenshot_1560459376.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/281069/attachments/Screenshot_1560459376.png)

Related work items: #700874