Merged PR 280661: BottomSheet V1
BottomSheet is used to display a list of menu items in a Drawer

This includes two implementations to set the list item data.
1. Via newInstance and parcelable data objects. Because the objects extends Parcelable they are able to be stored in Bundle allowing for the saving of this data on configuration change.
2. Via optional closures on data objects. Since the closures cannot be put into a Parcelable object they cannot be stored in the Bundle, therefore the BottomSheet is dismissed on configuration change.

In case a developer adds both data objects with closures and parcelable data objects, the data objects with closures win out and they are rendered in the BottomSheet.

Toast is used to demo the listeners and closures as the way DialogFragments are added to the Window interferes with the placement of Snackbar.

Additionally, refinements to the Drawer's dismissal are included in this PR.

![Screenshot_1560973104.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/280661/attachments/Screenshot_1560973104.png =400x)

Related work items: #726046