Merged PR 300659: BottomSheetDialog Fix for landscape automatic dismiss
When used in a persistent way, ie static instance or retained state, `BottomSheetDialog` dismisses after being shown in certain circumstances.

Repro:
1. Show `BottomSheetDialog` in portrait mode
2. Dismiss it
3. Rotate screen to landscape mode

This was caused by the `onSlide` `slideOffset` position being "-Infinity", triggering an automatic dismiss. Fix for this is to check if the `slideOffset` is greater than 0.

Retention of the dialog instance also created other issues after the dismiss bug was fixed which are addressed in the PR.
- rotation from landscape to portrait caused the animation start position of the dialog to be higher than it should've been. Fix for this is to call `requestLayout`.
- slide up animation was lost if the dialog is opened and the screen rotated. Fix for this is to set the `BottomSheetBehavior` state back to `STATE_COLLAPSED` so that the animation can be triggered for the dialog.

Also:
- fixed BottomSheetDialog and DrawerDialog demo memory leaks (this is dealt with in other PR, but that PR has other fixes as well)
- added static instance demo

Related work items: #753318