Merged PR 303830: PeoplePickerView: add ability to show hint and have a character threshold of 0
**Hint**
Originally we were hiding this because Outlook uses a label instead of hint to instruct users (but still had it for accessibility purposes). Now we have the option of showing the hint for this purpose.

**Character Threshold**
One of the `PeoplePickerTextView` base classes, `AutoCompleteTextView`, does not allow filtering (aka the dropdown won't show) when no characters have been entered in the field. Now if you set `characterThreshold` to 0, you'll get the dropdown of `availablePersonas` on focus.

| Hint shown on third field | The second field now has a 0 character threshold |
|-|-|
| ![Screenshot_1565363870.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/303830/attachments/Screenshot_1565363870.png)  | ![Screenshot_1565363863.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/303830/attachments/Screenshot_1565363863.png)  |

Related work items: #760301