Merged PR 317771: Replace old colors in new attributes
I double checked with the toolkit, and the button color was bumped to gray300 instead of the gray200 we had it mapped to.

`uifabricCalendarWeekHeadingWeekendTextColor` has an opacity in the toolkit, but I confirmed with design that gray200 is the intention, opacity was used as a shortcut to get the right color.

Related work items: #772661