Merged PR 341192: Searchbar: add style for action view type
`Searchbar` can now also be used as an action view in a menu, in addition to being used as an accessory view under a Toolbar.

To facilitate this, I added a new `isActionMenuView` boolean that helps define style changes needed in each mode. When true, I'm using the system back button so we get rid of ours.

Updated the "accessory view" demo to toggle between these new styles.

| Collapsed | Expanded |
| - | - |
| ![Screenshot_1572977663.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/341192/attachments/Screenshot_1572977663.png) | ![Screenshot_1572977665.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/32fa6338-45ea-42a0-aca0-484938e1962a/pullRequests/341192/attachments/Screenshot_1572977665.png) |

Related work items: #797437