Merged PR 343870: DateTimePicker: Unexpected End Time and Duration
This PR fixes a duration issue when picking an end time before a start time where both times are in close proximity to one another, ie 1 hour:

Repro steps:
1. Click "SELECT DATE TIME RANGE" button
2. Switch to time tab
3. Select 11 AM start time
4. For end time select one hour forward
5. Save => duration is still zero on both demo screen and in date picker

This is because a user is required to change the period to PM to get a proper duration.

The interaction improvement here is to automatically move the AM PM period according to the direction of time movement. This pattern is modeled after Outlook iOS' behavior

Related work items: #797775