Merged PR 344829: New Fluent Mobile Icons
I targeted all controls and demos with icons so if you see any old ones, let me know. I did not edit the hex values in the vector drawables because in most cases it's more efficient to use a tint for a drawable instead of duplicating drawables when we need new colors (a practice we've already started). Exception being special use cases like the Toolbar back button where we can't specify a tint. To be able to use a tint in the `BottomSheetItem`, I had to add a property for that.

I left the name the same as the icon repo's SVGs, but added the `ms_*` prefix when used in library controls. Exception being special use cases like the Toolbar back button where we need an extra suffix.

Related work items: #800468