Merged PR 394498: Add setter method to write default start of the week in shared preferences
~~See different approach in !394497~~

This is another approach to solve the same issue. In this case I am creating a setter to write a week start day in shared preferences. The downside of this approach is that we need changes in the calling app (see changed in the demo app).

There could be a another approach, which is combine both of them. Please, feel free to suggest and choose your preferred solution.

**Update:** I have combined both approaches in this PR, so it is now possible to write in shared preferences a different start of week, but in case there is no value written, the default from locale will be returned